pyproject.toml
[project]
name = "imageonpdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymupdf>=1.26.1",
    "pyyaml>=6.0.2",
]
